// extensions/quick-access/src/types/file-content-info.ts
export interface FileContentInfo {
  fileContent: string;
  name: string;
  where: string;
  sizeTitle: string;
  size: string;
  created: string;
  modified: string;
  lastOpened: string;
}
export const fileContentInfoInit = {
  fileContent: "",
  name: "",
  where: "",
  sizeTitle: "",
  size: "",
  created: "",
  modified: "",
  lastOpened: "",
};